requirements.txt
discord.py[speed]>=2.3.2
python-dotenv==1.0.1
git+https://github.com/Rapptz/asqlite
